Refactor Typescript and Python implementations into separate folders
