+++ .vscode/settings.json
@@ -7,6 +7,7 @@
   "typescript.tsdk": ".yarn/sdks/typescript/lib",
   "typescript.enablePromptUseWorkspaceTsdk": true,
   "cSpell.words": [
-    "gitly"
+    "gitly",
+    "konhi"
   ]
 }
