+++ src/npm/worker.ts
