+++ tsconfig.json
@@ -18,13 +18,13 @@
     "module": "esnext",
     "moduleResolution": "node",
     "outDir": "dist",
-    "rootDir": "src",
+    "rootDir": "./src/npm",
     "target": "ES2017",
     "lib": [
       "ES5",
       "ES6"
     ], /* Specify a set of bundled library declaration files that describe the target runtime environment. */
-    "baseUrl": "./src", /* Specify the base directory to resolve non-relative module names. */
+    "baseUrl": "./src/npm", /* Specify the base directory to resolve non-relative module names. */
     "types": [
       "node"
     ],
